Move test to tests/backend.test_register_allocation.py
